FrameとParaparaCanvasを統合し、階層構造を浅く
